Rename files to follow Google JS style guide
