+++ src/reducers/images.js
